Extract botfunctions class from main to new class Ui
