+++ src/main/java/Luna.java
@@ -1,6 +1,6 @@
 import java.io.InputStreamReader;
 import java.io.BufferedReader;
-import java.io.StringReader;
+//import java.io.StringReader;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -19,7 +19,8 @@ public class Luna {
 
         String NAME = "Luna"; // TENTATIVE
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        bot_functions.shifted_print(bot_functions.greetingString(NAME));
+//        private Ui ui = new Ui();
+        Ui.shifted_print(Ui.greetingString(NAME));
         boolean exitFlag = false;
         commandHints currentHint = commandHints.EXIT;
         ArrayList<list_Entry> user_list = new ArrayList<>();
@@ -36,24 +37,24 @@ public class Luna {
             if (input_command.equalsIgnoreCase("help")) {
                 switch(currentHint) {
                     case LIST:
-                        bot_functions.shifted_print("To display all the task in your list\n--> Type: list");
+                        Ui.shifted_print("To display all the task in your list\n--> Type: list");
                         break;
                     case TODO:
-                        bot_functions.shifted_print("To add a to-do task to your list\n--> Type: todo <task name>");
+                        Ui.shifted_print("To add a to-do task to your list\n--> Type: todo <task name>");
                         break;
                     case DEADLINE:
-                        bot_functions.shifted_print("To add a deadline task to your list\n--> Type: deadline <task name> <end date>");
+                        Ui.shifted_print("To add a deadline task to your list\n--> Type: deadline <task name> <end date>");
                         break;
                     case EVENT:
-                        bot_functions.shifted_print("To add a event task to your list\n--> Type: event <task name> <start date> <end date>");
+                        Ui.shifted_print("To add a event task to your list\n--> Type: event <task name> <start date> <end date>");
                         break;
                     case MARK:
-                        bot_functions.shifted_print("To mark a task as done\n--> Type: mark <list num>");
-                        bot_functions.shifted_print("To unmark a task as not done\n--> Type: unmark <list num>");
-                        bot_functions.shifted_print("To delete a task from the list\n--> Type: delete <list num>");
+                        Ui.shifted_print("To mark a task as done\n--> Type: mark <list num>");
+                        Ui.shifted_print("To unmark a task as not done\n--> Type: unmark <list num>");
+                        Ui.shifted_print("To delete a task from the list\n--> Type: delete <list num>");
                         break;
                     case EXIT:
-                        bot_functions.shifted_print("End the program :(\n--> Type: bye");
+                        Ui.shifted_print("End the program :(\n--> Type: bye");
                         break;
 
                 }
@@ -70,7 +71,7 @@ public class Luna {
             } else
             // IF EXIT
             if (input_command.equalsIgnoreCase("bye") || input_command.equalsIgnoreCase("exit")) {
-                bot_functions.shifted_print(bot_functions.signoffString());
+                Ui.shifted_print(Ui.signoffString());
                 exitFlag = true;
 
             // IF LIST IS REQUESTED
@@ -87,7 +88,7 @@ public class Luna {
                         text.append((i+1)).append(".").append(ent.toString()).append("\n");
                     }
                 }
-                bot_functions.shifted_print(text.toString());
+                Ui.shifted_print(text.toString());
 
             } else {
                 currentHint = commandHints.MARK;
@@ -96,27 +97,27 @@ public class Luna {
 
                 if (keys[0].equalsIgnoreCase("unmark") || keys[0].equalsIgnoreCase("mark") || keys[0].equalsIgnoreCase("delete")) {
                     if (Arrays.asList(keys).size() < 2) {
-                        bot_functions.shifted_print("Missing inputs");
-                    } else if (!bot_functions.isNumeric(keys[1])) {
-                        bot_functions.shifted_print("Please input a number after the command E.g (mark/unmark/delete) 3");
+                        Ui.shifted_print("Missing inputs");
+                    } else if (!Ui.isNumeric(keys[1])) {
+                        Ui.shifted_print("Please input a number after the command E.g (mark/unmark/delete) 3");
                     } else {
                         int pos = Integer.parseInt(keys[1]);
                         if (pos <= user_list.size() && pos > 0) {
                             list_Entry ent = user_list.get(pos-1);
                             if (keys[0].equalsIgnoreCase("delete")) {
                                 user_list.remove(pos-1);
-                                bot_functions.shifted_print("Deleting the following task:\n " + ent
+                                Ui.shifted_print("Deleting the following task:\n " + ent
                                         + "\nThere are " + user_list.size() + " tasks in the list");
                             }
                             else if (keys[0].equalsIgnoreCase("mark")) {
                                 ent.markEntry();
-                                bot_functions.shifted_print("Nice! I've marked this task as done:\n" + ent);
+                                Ui.shifted_print("Nice! I've marked this task as done:\n" + ent);
                             } else {
                                 ent.unmarkEntry();
-                                bot_functions.shifted_print("Nice! I've marked this task as not done yet:\n" + ent);
+                                Ui.shifted_print("Nice! I've marked this task as not done yet:\n" + ent);
                             }
                         } else {
-                            bot_functions.shifted_print("There are " + user_list.size() + " tasks in the list");
+                            Ui.shifted_print("There are " + user_list.size() + " tasks in the list");
                         }
                     }
                 } else {
@@ -155,9 +156,9 @@ public class Luna {
                         StringBuilder text = new StringBuilder();
                         String count_msg = "Oki! Added a task to the list:\n   " + ent + "\nNow you have " + user_list.size() + " tasks in the list.";
                         text.append(count_msg);
-                        bot_functions.shifted_print(text.toString());
+                        Ui.shifted_print(text.toString());
                     } else {
-                        bot_functions.shifted_print("Wrong format");
+                        Ui.shifted_print("Wrong format");
                     }
 
                 }
@@ -167,61 +168,6 @@ public class Luna {
     }
 
 
-    public static class bot_functions {
-        public static final String HORIZONTAL_LINE = "____________________________________________________________\n";
 
-        public static String greetingString(String name){
-            String msg = "";
-            msg += String.format("%30s", " （„• ֊ •„)♡\n");
-            msg += "heyo! my name is ✦" + name + "✦\n";
-            msg += "What would you like for me to do?\n";
-            return msg;
-        }
-
-        public static String sad_cat_img() {
-            return
-                    "⠀               />    フ\n" +
-                            "               | 　_  _|\n" +
-                            "           ___/` ミ＿_xノ\n" +
-                            "          /　　　     |\n" +
-                            "         /　 ヽ　     ﾉ\n" +
-                            "        │　    |　|　|\n" +
-                            "     /￣|　    |　|　|\n" +
-                            "    | (￣ ヽ＿ヽ_)__)\n" +
-                            "    ＼二二)⠀⠀\n";
-        }
-
-        public static String signoffString(){
-            String msg = "";
-            msg += "okay then, bye\n";
-            msg += sad_cat_img();
-
-            return msg;
-        }
-
-        public static void shifted_print(String text) throws Exception{
-            BufferedReader br = new BufferedReader(new StringReader(text));
-            String readText = br.readLine();
-            StringBuilder output = new StringBuilder();
-            output.append("      " + HORIZONTAL_LINE);
-            while (readText != null) {
-                output.append("      ").append(readText).append("\n");
-                readText = br.readLine();
-            }
-            output.append("      " + HORIZONTAL_LINE);
-            System.out.print(output);
-        }
-
-        // Reference from https://stackoverflow.com/questions/1102891/how-to-check-if-a-string-is-numeric-in-java
-        public static boolean isNumeric(String str) {
-            try {
-                Double.parseDouble(str);
-                return true;
-            } catch(NumberFormatException e){
-                return false;
-            }
-        }
-
-    }
 }
